Programatically invoke uvicorn for development server.
Invoke uvicorn from a python program so that we can configure it to
terminate connections when detects a file changed and it needs to be
reloaded.

Before this change uvicorn would hang open when a SSE event listener was
connected from the frontend, this would prevent a server reload until
that frontend connection was closed (with say a refresh).